imap-proto/parser: correct HighestModSeq type to u64
see https://github.com/djc/tokio-imap/pull/28#issuecomment-431669527

+++ imap-proto/src/types.rs
@@ -57,7 +57,7 @@ pub enum ResponseCode<'a> {
 
 #[derive(Debug, Eq, PartialEq)]
 pub enum StatusAttribute {
-    HighestModSeq(u32), // RFC 4551
+    HighestModSeq(u64), // RFC 4551
     Messages(u32),
     Recent(u32),
     UidNext(u32),
